moved statistics page to separate route
