+++ views/landing/landing.handlebars
@@ -6,7 +6,6 @@
     <div class="container p-0 m-0 mt-5 col-4">
         <ul class="list-group list-group-flush">
             <li class="list-group-item"><a href="/restaurants">View All Restaurants</a></li>
-            <li class="list-group-item"><a href="/restaurants/new">Create a New Restaurant</a></li>
             <li class="list-group-item"><a href="/statistics">Check Out Our Statistics Page</a></li>
         </ul>
     </div>
